Remove test function
I had left in a function that was only
meant to be there while testing the initial
version of the riemann client.
It is not in any way useful for any
application that uses the client, so
I have removed it.
Since it was an exported function this
does change the public API, but I hope this
does not cause any problems.